// types/leadSchema.ts
import { z } from 'zod';

export const leadSchema = z.object({
    ID: z.string(),
    TITLE: z.string(),
    TYPE_ID: z.string(),
    STAGE_ID: z.string(),
    PROBABILITY: z.string().nullable(),
    CURRENCY_ID: z.string(),
    OPPORTUNITY: z.string(),
    IS_MANUAL_OPPORTUNITY: z.string(),
    TAX_VALUE: z.string().nullable(),
    LEAD_ID: z.string(),
    COMPANY_ID: z.string(),
    CONTACT_ID: z.string(),
    QUOTE_ID: z.string().nullable(),
    BEGINDATE: z.string(),
    CLOSEDATE: z.string(),
    ASSIGNED_BY_ID: z.string(),
    CREATED_BY_ID: z.string(),
    MODIFY_BY_ID: z.string(),
    DATE_CREATE: z.string(),
    DATE_MODIFY: z.string(),
    OPENED: z.string(),
    CLOSED: z.string(),
    COMMENTS: z.string().nullable(),
    ADDITIONAL_INFO: z.string().nullable(),
    LOCATION_ID: z.string().nullable(),
    CATEGORY_ID: z.string(),
    STAGE_SEMANTIC_ID: z.string(),
    IS_NEW: z.string(),
    IS_RECURRING: z.string(),
    IS_RETURN_CUSTOMER: z.string(),
    IS_REPEATED_APPROACH: z.string(),
    SOURCE_ID: z.string(),
    SOURCE_DESCRIPTION: z.string().nullable(),
    ORIGINATOR_ID: z.string().nullable(),
    ORIGIN_ID: z.string().nullable(),
    MOVED_BY_ID: z.string(),
    MOVED_TIME: z.string(),
    LAST_ACTIVITY_TIME: z.string(),
    UTM_SOURCE: z.string().nullable(),
    UTM_MEDIUM: z.string().nullable(),
    UTM_CAMPAIGN: z.string().nullable(),
    UTM_CONTENT: z.string().nullable(),
    UTM_TERM: z.string().nullable(),
    PARENT_ID_144: z.string().nullable(),
    PARENT_ID_180: z.string().nullable(),
    LAST_ACTIVITY_BY: z.string(),
  });
  
  export type Lead = z.infer<typeof leadSchema>;